rename package patterns -> pattern
