+++ stix2/equivalence/__init__.py
@@ -3,7 +3,7 @@
 .. autosummary::
    :toctree: equivalence
 
-   patterns
+   pattern
    graph
    object
