+++ stix2/equivalence/pattern/compare/__init__.py
