+++ stix2/equivalence/pattern/transform/__init__.py
